created schedule-api

+++ flyway/sql/V2__Create_schedule_table.sql
@@ -1,5 +1,5 @@
 create table  if not exists SCHEDULE   (
     ID serial primary key,
-    user_id int,
-    schedule_id bigint,
+    user_id BIGINT,
+    schedule_id BIGINT,
 );
